+++ additional.h
@@ -1,7 +1,7 @@
 #ifndef ADDITIONAL_H_INCLUDED
 #define ADDITIONAL_H_INCLUDED
 
-int getmax(double *d);
+int getmax(float *d);
 void quickSort( int *a, int l, int r);
 
 #endif // ADDITIONAL_H_INCLUDED
